+++ apps/antalmanac/src/components/RightPane/CoursePane/SearchForm/LabeledInputs/LabeledTimePicker.tsx
@@ -1,10 +1,12 @@
-import { CustomInputBox } from '$components/RightPane/CoursePane/SearchForm/LabeledInputs/CustomInputBox';
-import { CustomInputLabel } from '$components/RightPane/CoursePane/SearchForm/LabeledInputs/CustomInputLabel';
-import { Box, TextField, type TextFieldProps } from '@mui/material';
-import { DesktopTimePicker, type TimePickerProps } from '@mui/x-date-pickers';
-import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
-import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
-import { forwardRef, useId, useState } from 'react';
+import { Skeleton } from '@mui/material';
+import type { TextFieldProps } from '@mui/material';
+import type { TimePickerProps } from '@mui/x-date-pickers';
+import dynamic from 'next/dynamic';
+
+const LabeledTimePickerContent = dynamic(() => import('./LabeledTimePickerContent'), {
+    ssr: false,
+    loading: () => <Skeleton variant="rounded" width="100%" height={40} sx={{ flex: 1 }} />,
+});
 
 interface LabeledTimePickerProps {
     label: string;
@@ -13,40 +15,4 @@ interface LabeledTimePickerProps {
     isAligned?: boolean;
 }
 
-export const LabeledTimePicker = ({ label, timePickerProps, textFieldProps, isAligned }: LabeledTimePickerProps) => {
-    const id = useId();
-    const [anchorEl, setAnchorEl] = useState<HTMLElement | null>(null);
-
-    const TimePickerTextField = forwardRef<HTMLInputElement, TextFieldProps>((params, ref) => (
-        <TextField size="small" variant="outlined" {...params} {...textFieldProps} id={id} inputRef={ref} />
-    ));
-    TimePickerTextField.displayName = 'TimePickerTextField';
-
-    return (
-        <LocalizationProvider dateAdapter={AdapterDateFns}>
-            <Box sx={{ display: 'flex', width: '100%', flex: 1 }}>
-                <CustomInputLabel label={label} id={id} isAligned={isAligned} />
-                <CustomInputBox
-                    boxProps={{
-                        ref: (node: HTMLElement | null) => {
-                            setAnchorEl(node);
-                        },
-                    }}
-                >
-                    <DesktopTimePicker
-                        enableAccessibleFieldDOMStructure={false}
-                        {...timePickerProps}
-                        slotProps={{
-                            popper: {
-                                anchorEl,
-                            },
-                        }}
-                        slots={{
-                            textField: TimePickerTextField,
-                        }}
-                    />
-                </CustomInputBox>
-            </Box>
-        </LocalizationProvider>
-    );
-};
+export const LabeledTimePicker = (props: LabeledTimePickerProps) => <LabeledTimePickerContent {...props} />;
