+++ apps/antalmanac/src/components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearch.tsx
@@ -2,29 +2,20 @@ import { AdvancedSearchFieldRow } from '$components/RightPane/CoursePane/SearchF
 import { BuildingField } from '$components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearchFields/BuildingField';
 import { DaysField } from '$components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearchFields/DaysField';
 import { DivisionField } from '$components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearchFields/DivisionField';
+import { EndTimeField } from '$components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearchFields/EndTimeField';
 import { ExcludeRestrictionsField } from '$components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearchFields/ExcludeRestrictionsField';
 import { ExcludeRoadmapField } from '$components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearchFields/ExcludeRoadmapField';
 import { FullCoursesField } from '$components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearchFields/FullCoursesField';
 import { InstructorField } from '$components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearchFields/InstructorField';
 import { OnlineField } from '$components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearchFields/OnlineField';
 import { RoomField } from '$components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearchFields/RoomField';
+import { StartTimeField } from '$components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearchFields/StartTimeField';
 import { UnitsField } from '$components/RightPane/CoursePane/SearchForm/ManualSearch/AdvancedSearch/AdvancedSearchFields/UnitsField';
 import { hasAdvancedParams } from '$components/RightPane/CoursePane/SearchParams/helpers';
 import { readAdvancedSearchParams } from '$components/RightPane/CoursePane/SearchParams/loaders';
 import { ExpandLess, ExpandMore } from '@mui/icons-material';
-import { Box, Button, Collapse, Skeleton, Typography } from '@mui/material';
-import dynamic from 'next/dynamic';
+import { Box, Button, Collapse, Typography } from '@mui/material';
 import { useState } from 'react';
-
-const StartTimeField = dynamic(
-    () => import('./AdvancedSearchFields/StartTimeField').then((m) => ({ default: m.StartTimeField })),
-    { ssr: false, loading: () => <Skeleton variant="rounded" width="100%" height={40} sx={{ flex: 1 }} /> }
-);
-
-const EndTimeField = dynamic(
-    () => import('./AdvancedSearchFields/EndTimeField').then((m) => ({ default: m.EndTimeField })),
-    { ssr: false, loading: () => <Skeleton variant="rounded" width="100%" height={40} sx={{ flex: 1 }} /> }
-);
 
 export function AdvancedSearch() {
     const [expanded, setExpanded] = useState(() => hasAdvancedParams(readAdvancedSearchParams()));
